deployment/requirements.txt
pandas
numpy
tensorflow
re
nltk
tensorflow_hub
tensorflow.keras.models
string
streamlit
matplotlib.pyplot